Modified ExitCode type
